api/pyproject.toml
[project]
name = "mcp-linker-api"
version = "1.0.1"
description = "MCP-Linker Backend API with Supabase Authentication integration."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi[all]>=0.100.0",
    "fastapi-cache2[redis]>=0.2.1",
    "fastapi-limiter>=0.1.5",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyjwt>=2.10.1",
    "python-jose>=3.4.0",
    "redis>=4.6.0",
    "uvicorn[standard]>=0.23.0",
    "prisma>=0.15.0",
]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true



[dependency-groups]
dev = ["faker>=37.1.0", "scrapy>=2.13.0", "tomlkit>=0.13.2"]
